[address] Refactor to be more clear
