+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the Aptos Go SDK will be captured in this file. This chan
 adheres to the format set out by [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Unreleased
+- Refactor address internal packages to be more clear
 
 # v1.10.0 (6/20/2025)
 - [`Feature`] Add orderless transaction support
