+++ internal/types/accountAddress.go
@@ -1,13 +1,20 @@
 package types
 
 import (
+	"encoding/hex"
 	"encoding/json"
+	"errors"
 	"fmt"
+	"strings"
 
 	"github.com/aptos-labs/aptos-go-sdk/bcs"
 	"github.com/aptos-labs/aptos-go-sdk/crypto"
 	"github.com/aptos-labs/aptos-go-sdk/internal/util"
 )
+
+// -----
+// Note that all of these are re-exported, and are only internal to prevent circular dependencies
+// -----
 
 // AccountAddress a 32-byte representation of an on-chain address
 //
@@ -54,6 +61,8 @@ func (aa *AccountAddress) IsSpecial() bool {
 }
 
 // String Returns the canonical string representation of the [AccountAddress]
+//
+// These are AIP-40 compliant.
 //
 // Please use [AccountAddress.StringLong] for all indexer queries.
 func (aa *AccountAddress) String() string {
@@ -132,4 +141,61 @@ func (aa *AccountAddress) DerivedAddress(seed []byte, typeByte uint8) AccountAdd
 	authKey := aa.AuthKey()
 	authKey.FromBytesAndScheme(append(authKey[:], seed...), typeByte)
 	return AccountAddress(authKey[:])
+}
+
+// ErrAddressMissing0x is returned when an AccountAddress is missing the leading 0x
+var ErrAddressMissing0x = errors.New("AccountAddress missing 0x")
+
+// ErrAddressTooShort is returned when an AccountAddress is too short
+var ErrAddressTooShort = errors.New("AccountAddress too short")
+
+// ErrAddressTooLong is returned when an AccountAddress is too long
+var ErrAddressTooLong = errors.New("AccountAddress too long")
+
+// ParseStringRelaxed parses a string into an AccountAddress
+// TODO: add strict mode checking
+func (aa *AccountAddress) ParseStringRelaxed(x string) error {
+	x = strings.TrimPrefix(x, "0x")
+	if len(x) < 1 {
+		return ErrAddressTooShort
+	}
+	if len(x) > 64 {
+		return ErrAddressTooLong
+	}
+	if len(x)%2 != 0 {
+		x = "0" + x
+	}
+	bytes, err := hex.DecodeString(x)
+	if err != nil {
+		return err
+	}
+	// zero-prefix/right-align what bytes we got
+	copy((*aa)[32-len(bytes):], bytes)
+
+	return nil
+}
+
+// ParseStringWithPrefixRelaxed parses a string into an AccountAddress
+func (aa *AccountAddress) ParseStringWithPrefixRelaxed(x string) error {
+	if !strings.HasPrefix(x, "0x") {
+		return ErrAddressMissing0x
+	}
+	x = x[2:]
+	if len(x) < 1 {
+		return ErrAddressTooShort
+	}
+	if len(x) > 64 {
+		return ErrAddressTooLong
+	}
+	if len(x)%2 != 0 {
+		x = "0" + x
+	}
+	bytes, err := hex.DecodeString(x)
+	if err != nil {
+		return err
+	}
+	// zero-prefix/right-align what bytes we got
+	copy((*aa)[32-len(bytes):], bytes)
+
+	return nil
 }
